examples: address ErrorProne and warnings. (#2433)
Discovered when importing grpc to internal repo:

1. ErrorProne: Compound assignments from floating point to integral
types hide dangerous casts.

2. found raw type: io.grpc.ManagedChannelBuilder missing type arguments
  for generic class io.grpc.ManagedChannelBuilder<T>

+++ examples/src/main/java/io/grpc/examples/routeguide/RouteGuideServer.java
@@ -290,7 +290,7 @@ public class RouteGuideServer {
      * @param end The end point
      * @return The distance between the points in meters
      */
-    private static double calcDistance(Point start, Point end) {
+    private static int calcDistance(Point start, Point end) {
       double lat1 = RouteGuideUtil.getLatitude(start);
       double lat2 = RouteGuideUtil.getLatitude(end);
       double lon1 = RouteGuideUtil.getLongitude(start);
@@ -305,7 +305,7 @@ public class RouteGuideServer {
           + cos(phi1) * cos(phi2) * sin(deltaLambda / 2) * sin(deltaLambda / 2);
       double c = 2 * atan2(sqrt(a), sqrt(1 - a));
 
-      return r * c;
+      return (int) r * c;
     }
   }
 }
